Add first basic envelope implementation

+++ src/envelope.rs
@@ -7,7 +7,7 @@ pub enum EnvStage {
 }
 
 pub struct Envelope {
-    pub duration: f32,
+    pub duration: u32,
     pub phase: f64,
     pub hz: f64,
 }
